Update README.md
I added a few  of the suggestions and I'll continue working on the readme(I wanted to improve the readme gradually with your suggestions and look at other examples). Thanks so much for the help and patience. This is actually my first time contributing to another persons project. 
Thank you again!